+++ src/App.js
@@ -1,8 +1,6 @@
-import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Calculator from './components/Calculator';
-import ButtonBox from './components/ButtonBox';
-import Screen from './components/Screen';
+
 import Quotes from './components/Quotes';
 import Home from './components/Home';
 import NotFoundPage from './components/NotFound';
@@ -10,13 +8,6 @@ import './App.css';
 import Navbar from './components/Navbar';
 
 function App() {
-  const [displayValue, setDisplayValue] = React.useState('0');
-
-  const handleButtonClick = (calculatorData) => {
-    const { total, next } = calculatorData;
-    setDisplayValue(next || total || '0');
-  };
-
   return (
     <BrowserRouter>
       <Navbar />
@@ -24,10 +15,7 @@ function App() {
         <Route
           path="/"
           element={(
-            <Calculator>
-              <Screen value={displayValue} />
-              <ButtonBox onButtonClick={handleButtonClick} />
-            </Calculator>
+            <Calculator />
           )}
         />
         <Route path="/quotes" element={<Quotes />} />
